changed function names and fixed funciton to long
